+++ src/geant4/app/exampleN01.cc
@@ -15,7 +15,18 @@
 #include "G4Timer.hh"
 #include "G4UIExecutive.hh"
 #include "G4UImanager.hh"
+
+#ifdef GEANT4_GT_10_6
+#include "G4UIExecutive.hh"
 #include "G4VisExecutive.hh"
+#else
+#ifdef G4VIS_USE
+#include "G4VisExecutive.hh"
+#endif
+#ifdef G4UI_USE
+#include "G4UIExecutive.hh"
+#endif
+#endif
 
 #ifndef uwuw_hpp
 #define uwuw_hpp 1